requirements.txt
fastapi[all]absl-py==2.0.0
fastapi==0.110.0
openai==1.14.1
openapi-schema-pydantic==1.2.4
pandas==1.5.3
pinecone-client==3.1.0
uvicorn==0.28.0